requirements.txt
aiosmtpd==1.4.6
atpublic==6.0.2
attrs==25.4.0
dnspython==2.7.0
python-dotenv==1.2.1
